fix(data): entity attribute defaults, subclass mob effect particles, VarInt packet ids
- entity_attributes.json + extract_entity_attributes.py resolve Paper's
  DefaultAttributes chains per EntityType; build.rs emits
  EntityKind::is_living() and EntityKind::default_attribute(attr)
- extract_mob_effects.py follows new XxxMobEffect(...) subclasses for the
  super(category, color, ParticleTypes.X) particle (wind_charged, weaving,
  oozing, infested)
- Java float literals widen to f64 like Java; serde_json float_roundtrip so
  the generated constants match the JSON bit for bit
- ModifierOperation encodes as a VarInt, PlayPacket sorted by id,
  packet_id attribute parsed as i32

+++ void-codec-macros/src/decode.rs
@@ -116,7 +116,7 @@ pub fn derive_decode(input: &DeriveInput) -> Result<proc_macro2::TokenStream> {
                         let variant_attrs = parse_variant_attrs(&v.attrs)?;
 
                         let packet_id = match variant_attrs.packet_id {
-                            Some(id) => id as i32,
+                            Some(id) => id,
                             None => {
                                 return Err(Error::new_spanned(
                                 v,
